WebGL frontend via JavaScript and wasm-bindgen

See merge request remram44/vigilant-steel!1

+++ benches/simple.rs
@@ -11,7 +11,7 @@ fn flying_asteroids(b: &mut Bencher) {
     b.bytes = 12;
     b.iter(|| {
         let mut game = Game::new_standalone();
-        const TIME_STEP: f64 = 0.040;
+        const TIME_STEP: f32 = 0.040;
         // Simulate 20 seconds of game time
         for frame in 0..500 {
             game.update(TIME_STEP);
